--- springRepo/gamepound/src/main/java/com/gamepound/app/review/service/ReviewServiceLKM.java
@@ -1,22 +1,0 @@
-package com.gamepound.app.review.service;
-
-import org.mybatis.spring.SqlSessionTemplate;
-import org.springframework.stereotype.Service;
-
-import com.gamepound.app.review.dao.ReviewDaoLKM;
-import com.gamepound.app.review.vo.ReviewVo;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class ReviewServiceLKM {
-
-	private final SqlSessionTemplate sst;
-	private final ReviewDaoLKM dao;
-	
-	// 리뷰 작성
-	public int write(ReviewVo vo) {	
-		return dao.write(sst, vo);
-	}
-}
