+++ springRepo/gamepound/src/main/java/com/gamepound/app/userpage/service/UserPageService.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import org.mybatis.spring.SqlSessionTemplate;
 import org.springframework.stereotype.Service;
 
+import com.gamepound.app.project.vo.ProjectVo;
 import com.gamepound.app.review.vo.ReviewStatVo;
 import com.gamepound.app.review.vo.ReviewVo;
 import com.gamepound.app.userpage.dao.UserPageDao;
@@ -34,6 +35,18 @@ public class UserPageService {
 		map.put("statVo", statVo);
 		
 		return map;
+	}
+
+	// 리뷰 작성
+	public int write(ReviewVo vo) {	
+		return dao.write(sst, vo);
+	}
+	
+	// 유저페이지 - 내가 올린 프로젝트 목록 조회
+	public List<ProjectVo> listMyProjects(String memberNo) {
+		List<ProjectVo> myProjectList = dao.listMyProjects(sst, memberNo);
+
+		return myProjectList;
 	}
 	
 	// 유저페이지 -
